Remove vertices when all connected triangles are filled.
The implementation is really just lowering the vertex and its number to the back of the canvas (so that it is covered by triangles). When undo is used in addition to making every triangle connected to that edge transparent, it also raises every vertex connected to those triangles to the top of the canvas.
Closes #7